ref: 하단 내비게이션 삭제 및 플로팅 버튼 추가
