--- app/(tabs)/home.tsx
@@ -1,7 +1,0 @@
-import { Redirect } from "expo-router";
-
-const Home = () => {
-  return <Redirect href={"/my"} />;
-};
-
-export default Home;
