+++ app/[teamId]/index.tsx
